+++ composeApp/src/iosMain/kotlin/br/com/navalbattle/data/Clock.ios.kt
@@ -1,0 +1,6 @@
+package br.com.navalbattle.data
+
+import platform.Foundation.NSDate
+import platform.Foundation.timeIntervalSince1970
+
+actual fun nowMillis(): Long = (NSDate().timeIntervalSince1970 * 1000).toLong()